adding events input to recipe modules and refactoring
